requirements.txt
streamlit
pandas
requests
plotly
geopy
numpy
mysql-connector-python
fastapi
uvicorn
python-dotenv